DRY-up `Documents`
The exact same code was being declared in both the
`CallForEvidencePresenter` and `ConsultationPresenter`.

Therefore DRY-ing up to reduce the code duplication.
